Enabled "treat warnings as errors" and fixed existing warnings.

+++ PerfFlow/src/sampling/debugclient/ComCallStack.cpp
@@ -13,7 +13,7 @@ bool PerfFlow::ComCallStack::sample(const ComPtr<IDebugControl>& debugControl)
 {
 	_frameCount = 0;
 	ULONG filled;
-	if (FAILED(debugControl->GetStackTrace(0, 0, 0, _frames.data(), _frames.size(), &filled)))
+	if (FAILED(debugControl->GetStackTrace(0, 0, 0, _frames.data(), static_cast<ULONG>(_frames.size()), &filled)))
 		return false;
 
 	_frameCount = filled;
